+++ src/main/kotlin/mx/com/team9/activities/Gasto.kt
@@ -1,18 +1,16 @@
 package mx.com.team9.activities
 
-import mx.com.team9.dataclase.Categoria
-import mx.com.team9.models.Cuenta
-import java.time.LocalDateTime
+import mx.com.team9.models.Categoria
 
 /**
  * Clase que representa un retiro de una cuenta
  */
-class Retiro(
+class Gasto(
     override val monto: Double,
     override val descripcion: String,
     override val categoria: Categoria,
     override val cuentaOrigen: Cuenta
-) : Transaccion() {
+) : Movimiento() {
 
 
     init {
