+++ src/main/kotlin/mx/com/team9/activities/Ingreso.kt
@@ -1,16 +1,13 @@
 package mx.com.team9.activities
 
-import mx.com.team9.dataclase.Categoria
-import mx.com.team9.models.Cuenta
-import mx.com.team9.utils.TipoMovimiento
-import java.time.LocalDateTime
+import mx.com.team9.models.Categoria
 
-class Deposito(
+class Ingreso(
     override val monto: Double,
     override val descripcion: String,
     override val categoria: Categoria,
     override val cuentaOrigen: Cuenta
-) : Transaccion() {
+) : Movimiento() {
     init {
         cuentaOrigen.saldo += monto
         logTransaccion()
